Fix validation order causing uncaught error in earlier code and refactor validation
